[OT-348][FIX]: shortsApi.ts 안 patch인 updateShortsApi 별도로 분리
