+++ src/entities/shorts/apis/index.ts
@@ -1,2 +1,3 @@
 export * from "./getShortsApi";
 export * from "./shortsApi";
+export * from "./updateShortsApi";
